web: Common: Renamed folder people-selector to PeopleSelector
