+++ web/ASC.Web.Common/src/components/PeopleSelector/PeopleSelector.js
