+++ web/ASC.Web.Common/src/components/PeopleSelector/PeopleSelector.stories.js
