+++ web/ASC.Web.Common/src/components/PeopleSelector/index.js
